fix : 🚑 jwt 토큰에 대한 학습, 토큰값 관련 변경 로직. boolean 관계.
1. jwt 토큰 로직 관련 36번째 JwtAuthenticationProvider 참고.

2. !역으로 되어 있던 관계, -> !을 푼 관계로 변경.(모두 적용 완료)

3. Aes256Utils 사용하지 않아 삭제.

4.Accommodation 관련 controller/service 필요하지 않아 삭제

5.from Dto에 생성해 진행

6. UserController, UserService로 통일.(toEntity) 형태로 respose변경 예정

7. 현재 toEntity가 잘 되지 않아 헤매는 중 ㅠ

BookController 위치

8. ResponseCode SUCCESS 외에는 여기서 쓰는 게 없기 때문에, SUCCESS 만 사용 가능하게 남김

9. repository 에서 _를 이용해 해당 테이블의 원하는 값을 가져오는 방식으로 repository 로직 변경(멘토링 내용 반영)

10. Data 삭제 (ResponseAPI -> toString() 구문도 사용하지 않아 삭제 )
