+++ src/main/java/com/zerobase/yogizogi/accommodation/repository/PriceRepository.java
@@ -1,12 +1,11 @@
 package com.zerobase.yogizogi.accommodation.repository;
 
 import com.zerobase.yogizogi.accommodation.domain.entity.Price;
-import com.zerobase.yogizogi.accommodation.domain.entity.Room;
 import java.time.LocalDate;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface PriceRepository extends JpaRepository<Price, Long> {
-   Price findAllByRoomAndDate(Room room, LocalDate date);
+   Price findAllByRoom_IdAndDate(Long roomId, LocalDate date);
 }
