+++ src/main/java/com/zerobase/yogizogi/book/dto/BookResultDto.java
@@ -1,5 +1,6 @@
 package com.zerobase.yogizogi.book.dto;
 
+import com.zerobase.yogizogi.book.domain.entity.Book;
 import java.time.LocalDate;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -20,7 +21,19 @@ public class BookResultDto {
     String picUrl;
     LocalDate checkInDate;
     LocalDate checkOutDate;
-    Integer price;
-    Double score;
+    Integer payAmount;
+    Double rate;
     Boolean reviewRegistered;
+    //정적 팩터리 메서드
+    public static BookResultDto from(Book book){
+        return BookResultDto.builder()
+            .bookName(book.getBookName())
+            .payAmount(book.getPayAmount())
+            .rate(book.getAccommodation().getRate())
+            .checkInDate(book.getCheckInDate())
+            .checkOutDate(book.getCheckOutDate())
+            .picUrl(book.getAccommodation().getPicUrl())
+            .reviewRegistered(book.getReviewRegistered())
+            .build();
+    }
 }
