+++ src/main/java/com/zerobase/yogizogi/global/ResponseCode.java
@@ -8,8 +8,15 @@ import org.springframework.http.HttpStatus;
 @Getter
 public enum ResponseCode implements Code{
 
-    RESPONSE_SUCCESS(HttpStatus.OK, "SUCCESS"),
-    RESPONSE_FAIL(HttpStatus.BAD_REQUEST, "FAIL");
-    private final HttpStatus status;
-    private final String msg;
+    RESPONSE_SUCCESS {
+        @Override
+        public HttpStatus getStatus() {
+            return HttpStatus.OK;
+        }
+
+        @Override
+        public String getMsg() {
+            return "SUCCESS";
+        }
+    }
 }
