+++ src/main/java/com/zerobase/yogizogi/global/config/CorsConfig.java
@@ -12,7 +12,7 @@ public class CorsConfig implements WebMvcConfigurer {
     public void addCorsMappings(CorsRegistry registry) {
         registry.addMapping("/**")
             .allowedOrigins("*")
-            .allowedMethods("GET","POST","PUT","DELETE")
+            .allowedMethods("GET","POST","PATCH","DELETE")
             .allowedHeaders("*");
     }
 }
