+++ src/main/java/com/zerobase/yogizogi/review/repository/ReviewRepository.java
@@ -1,9 +1,6 @@
 package com.zerobase.yogizogi.review.repository;
 
-import com.zerobase.yogizogi.accommodation.domain.entity.Accommodation;
 import com.zerobase.yogizogi.review.domain.entity.Review;
-import java.util.List;
-import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -11,6 +8,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface ReviewRepository extends JpaRepository<Review, Long> {
-    Page<Review> findAllByAccommodation(Accommodation accommodation, Pageable pageable);
+    Page<Review> findAllByAccommodation_Id(Long accommodationId, Pageable pageable);
 
 }
